requirements.txt
torch
torchvision
pillow
pyyaml
tqdm
numpy
